% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_main_labels_factors.R
\name{add_main_labels_factors}
\alias{add_main_labels_factors}
\title{Add labels and factor levels to the merged data prior to deriving new variables.}
\usage{
add_main_labels_factors(data)
}
\arguments{
\item{data}{A data frame containing VMAC variables.}
}
\value{
\code{data} with added labels and factors.
}
\description{
Add labels and factor levels to the merged data prior to deriving new variables.
}
