% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/call_function.R
\name{call_function}
\alias{call_function}
\title{Call a function on a data set and print a table of added variables.}
\usage{
call_function(function_name, data, ...)
}
\arguments{
\item{function_name}{The name of a sourced function.}

\item{data}{A data frame containing VMAC variables.}

\item{...}{Additional arguments to be passed on to the function in \code{function_name}.}
}
\value{
\code{data} modified by \code{function_name} along with a data frame listing added variables.
}
\description{
Call a function on a data set and print a table of added variables.
}
