% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_shared_vars.R
\name{check_shared_vars}
\alias{check_shared_vars}
\title{Check if two data frames both contain variables with the same name. If so, a warning is printed and merging the two data frames will result in variables with suffixes ".x" and ".y".}
\usage{
check_shared_vars(data1, data2, merge.by)
}
\arguments{
\item{data1}{A data frame.}

\item{data2}{A second data frame.}

\item{merge.by}{A variable name or a list of variable names to exclude in the check.}
}
\value{
If any shared variables are found, a warning with the list of shared variables.
}
\description{
Check if two data frames both contain variables with the same name. If so, a warning is printed and merging the two data frames will result in variables with suffixes ".x" and ".y".
}
