% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwise_test.R
\name{chisq.post.hoc}
\alias{chisq.post.hoc}
\title{Compute p-values for pairwise comparisons for continuous and categorical variables by group.}
\usage{
chisq.post.hoc(tbl, test = c("fisher.test"), popsInRows = TRUE,
  control = c("fdr", "BH", "BY", "bonferroni", "holm", "hochberg",
  "hommel", "none"), digits = 4, ...)
}
\arguments{
\item{data}{A data frame containing VMAC variables.}

\item{variable}{A vector of continuous or categorical variables.}

\item{group}{A categorical variable by which the pariwise comparisons will be made.}

\item{continuous}{An integer value indicating the minimum number of unique values for \code{variable} to be considered continuous.}

\item{p.adj}{Method for adjusting p values (see p.adjust).}

\item{datatable}{A logical value indicating whether to return a formatted datatable (TRUE) or an unformatted dataframe (FALSE).}
}
\value{
A dataframe containing p-values of pairwise comparisons.
}
\description{
Compute p-values for pairwise comparisons for continuous and categorical variables by group.
}
