% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clear_labels.R
\name{clear_labels}
\alias{clear_labels}
\title{Remove Hmisc labels from a labelled data frame and return the unlabelled data frame along with a list of the labels.}
\usage{
clear_labels(data)
}
\arguments{
\item{data}{A labelled data frame or list}
}
\value{
\code{data} without labels and a named character vector containing the labels with object name \code{{data}_labels}.
}
\description{
Remove Hmisc labels from a labelled data frame and return the unlabelled data frame along with a list of the labels.
}
\examples{
\dontrun{
mydat.nolabels.df <- clear_labels(mydat)
}
}
