% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coerce_and_stack.R
\name{coerce_and_stack}
\alias{coerce_and_stack}
\title{Check whether variables with the same name in two data frames have different storage modes. If so, the numeric vector is coerced to character. The rows of the new data frame are then added to the old data frame.}
\usage{
coerce_and_stack(new_data, old_data)
}
\arguments{
\item{new_data}{A data frame to be bound to another data frame by rows.}

\item{old_data}{A data frame to which \code{new_data} will be bound by rows.}
}
\value{
A data frame with \code{new_data} bound by rows to \code{old_data}.
}
\description{
Check whether variables with the same name in two data frames have different storage modes. If so, the numeric vector is coerced to character. The rows of the new data frame are then added to the old data frame.
}
