% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_data.R
\name{compare_data}
\alias{compare_data}
\title{Compare additions, modifications, and deletions between two data frames with \code{daff::diff_data}.}
\usage{
compare_data(data1, data2, file = paste0("data_comparison_",
  data.merge.bh.file, ".html"))
}
\arguments{
\item{data1}{A data frame.}

\item{data2}{A second data frame.}

\item{file}{A path for where the output is saved.}
}
\value{
HTML formatted output summarizing differences between \code{data1} and \code{data2}.
}
\description{
Compare additions, modifications, and deletions between two data frames with \code{daff::diff_data}.
}
\seealso{
Other comparison functions: \code{\link{comparison}}
}
\concept{comparison functions}
