% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparison.R
\name{comparison}
\alias{comparison}
\alias{compare_variables}
\alias{compare_ids}
\alias{compare_factors}
\title{A variety of data comparison functions to use as checks prior to a destructive merge.}
\usage{
compare_variables(new_data, old_data)

compare_ids(new_data, old_data)

compare_factors(new_data, old_data)
}
\arguments{
\item{new_data}{A data frame.}

\item{old_data}{Another data frame.}
}
\value{
A tabulated list of differences between \code{new_data} and \code{old_data}.
}
\description{
A variety of data comparison functions to use as checks prior to a destructive merge.
}
\section{Functions}{
\itemize{
\item \code{compare_variables}: Outputs a tabulated list of differences in variable names between \code{new_data} and \code{old_data}.

\item \code{compare_ids}: Outputs a tabulated list of differences in MAP IDs between \code{new_data} and \code{old_data}.

\item \code{compare_factors}: Outputs a tabulated list of differences in factor levels between \code{new_data} and \code{old_data}.
}}

\seealso{
Other comparison functions: \code{\link{compare_data}}
}
\concept{comparison functions}
