% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_dates.R
\name{convert_dates}
\alias{convert_dates}
\title{Convert date variables to "Date" class.}
\usage{
convert_dates(data)
}
\arguments{
\item{data}{A data frame containing VMAC variables.}
}
\value{
\code{data} with date variables converted to "Date" class.
}
\description{
Convert date variables to "Date" class.
}
