% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_freeze.R
\name{data_freeze}
\alias{data_freeze}
\title{Freeze data by downloading it from REDCap. Static data sets are also added to the list object. The API tokens must be unlocked from a secure, encrypted vault.}
\usage{
data_freeze(box.dir = file.path("~", "box"), quarterly.download,
  tokens.list, redcap.api.uri = "https://redcap.vanderbilt.edu/api/",
  save = TRUE, return = TRUE)
}
\arguments{
\item{box.dir}{User path to Box home directory (parent directory of "VMAC BIOSTAT").}

\item{quarterly.download}{A logical value indicating whether this is a regularly scheduled download or if it is off-cycle (\code{scheduled download = TRUE}; \code{interim download = FALSE}). This variable determines the data freeze save location.}

\item{tokens.list}{The list of databases and API tokens to download from REDCap.}

\item{redcap.api.uri}{The URI for the REDCap API}

\item{save}{A logical value indicating whether to save the output as an RDS file in the default directory ("rawData" or "rawData/temp", depending on the value of \code{quarterly.download})}

\item{return}{A logical value indicating whether to return the output.}
}
\value{
A comprehensive list object containing all the raw data necessary to perform a data merge.
}
\description{
Freeze data by downloading it from REDCap. Static data sets are also added to the list object. The API tokens must be unlocked from a secure, encrypted vault.
}
\examples{
\dontrun{
MAPfreeze_interim.list <- data_freeze(quarterly.download = FALSE, save = TRUE, return = TRUE)
}
}
