% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derive_abp_qc.R
\name{derive_abp_qc}
\alias{derive_abp_qc}
\title{Derive and add ABP QC derived variables to the merged data set.}
\usage{
derive_abp_qc(data)
}
\arguments{
\item{data}{A data frame containing VMAC variables.}
}
\value{
\code{data} with added ABP QC derived variables.
}
\description{
Derive and add ABP QC derived variables to the merged data set.
}
