% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derive_AD_signature.R
\name{derive_AD_signature}
\alias{derive_AD_signature}
\title{Derive, label, and add AD signature McEvoy and Schwarz variables to the merged data set.}
\usage{
derive_AD_signature(data)
}
\arguments{
\item{data}{A data frame containing VMAC variables.}
}
\value{
\code{data} with added AD signature variables.
}
\description{
Derive, label, and add AD signature McEvoy and Schwarz variables to the merged data set.
}
