% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derive_addendum.R
\name{derive_addendum}
\alias{derive_addendum}
\title{Derive, label, and add addendum variables to the merged data set.}
\usage{
derive_addendum(data)
}
\arguments{
\item{data}{A data frame containing VMAC variables.}
}
\value{
\code{data} with added addendum variables.
}
\description{
Derive, label, and add addendum variables to the merged data set.
}
