% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derive_apoe.R
\name{derive_apoe}
\alias{derive_apoe}
\title{Derive, label, and add ApoE variables to the merged data set.}
\usage{
derive_apoe(data)
}
\arguments{
\item{data}{A data frame containing VMAC variables.}
}
\value{
\code{data} with added ApoE variables.
}
\description{
Derive, label, and add ApoE variables to the merged data set.
}
