% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derive_automated_3T.R
\name{derive_automated_3T}
\alias{derive_automated_3T}
\title{Derive, label, and add automated 3T variables to the merged data set.}
\usage{
derive_automated_3T(data)
}
\arguments{
\item{data}{A data frame containing VMAC variables.}
}
\value{
\code{data} with added automated 3T variables.
}
\description{
Derive, label, and add automated 3T variables to the merged data set.
}
