% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derive_biomarkers.R
\name{derive_biomarkers}
\alias{derive_biomarkers}
\title{Derive, label, and add biomarker variables to the merged data set.}
\usage{
derive_biomarkers(data)
}
\arguments{
\item{data}{A data frame containing VMAC variables.}
}
\value{
\code{data} with added biomarker variables.
}
\description{
Derive, label, and add biomarker variables to the merged data set.
}
