% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derive_blood.R
\name{derive_blood}
\alias{derive_blood}
\title{Derive, convert, label, and add blood variables to the merged data set.}
\usage{
derive_blood(data)
}
\arguments{
\item{data}{A data frame containing VMAC variables.}
}
\value{
\code{data} with added blood variables.
}
\description{
Derive, convert, label, and add blood variables to the merged data set.
}
