% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derive_cardiac_structure.R
\name{derive_cardiac_structure}
\alias{derive_cardiac_structure}
\title{Derive, label, and add cardiac structure variables to the merged data set.}
\usage{
derive_cardiac_structure(data)
}
\arguments{
\item{data}{A data frame containing VMAC variables.}
}
\value{
\code{data} with added cardiac structure variables.
}
\description{
Derive, label, and add cardiac structure variables to the merged data set.
}
