% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derive_cdr_conversion.R
\name{derive_cdr_conversion}
\alias{derive_cdr_conversion}
\title{Derive CDR conversion data.}
\usage{
derive_cdr_conversion(data)
}
\arguments{
\item{data}{A data frame containing VMAP data.}
}
\value{
A processed version of \code{data}.
}
\description{
Derive CDR conversion data.
}
