% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derive_cognitive_complaint.R
\name{derive_cognitive_complaint}
\alias{derive_cognitive_complaint}
\title{Derive, label, and add cognitive complaint, FCADL, and FAQ variables to the merged data set.}
\usage{
derive_cognitive_complaint(data)
}
\arguments{
\item{data}{A data frame containing VMAC variables.}
}
\value{
\code{data} with added cognitive complaint, FCADL, and FAQ variables.
}
\description{
Derive, label, and add cognitive complaint, FCADL, and FAQ variables to the merged data set.
}
