% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derive_cow_variants.R
\name{derive_cow_variants}
\alias{derive_cow_variants}
\title{Derive, label, and add Circle of Willis variables to the merged data set.}
\usage{
derive_cow_variants(data)
}
\arguments{
\item{data}{A data frame containing VMAC variables.}
}
\value{
\code{data} with added Circle of Willis variables.
}
\description{
Derive, label, and add Circle of Willis variables to the merged data set.
}
