% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derive_csf.R
\name{derive_csf}
\alias{derive_csf}
\title{Derive, convert, label, and add CSF variables to the merged data set.}
\usage{
derive_csf(data)
}
\arguments{
\item{data}{A data frame containing VMAC variables.}
}
\value{
\code{data} with added CSF variables.
}
\description{
Derive, convert, label, and add CSF variables to the merged data set.
}
