% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derive_dates.R
\name{derive_dates}
\alias{derive_dates}
\title{Derive, label, and add date variables to the merged data set.}
\usage{
derive_dates(data)
}
\arguments{
\item{data}{A data frame containing VMAC variables.}
}
\value{
\code{data} with added date variables.
}
\description{
Derive, label, and add date variables to the merged data set.
}
