% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derive_demographics.R
\name{derive_demographics}
\alias{derive_demographics}
\title{Derive, label, and add demographic variables to the merged data set.}
\usage{
derive_demographics(data)
}
\arguments{
\item{data}{A data frame containing VMAC variables.}
}
\value{
\code{data} with added demographic variables.
}
\description{
Derive, label, and add demographic variables to the merged data set.
}
