% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derive_depression.R
\name{derive_depression}
\alias{derive_depression}
\title{Derive, label, and add depression variables to the merged data set.}
\usage{
derive_depression(data)
}
\arguments{
\item{data}{A data frame containing VMAC variables.}
}
\value{
\code{data} with added depression variables.
}
\description{
Derive, label, and add depression variables to the merged data set.
}
