% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derive_diagnosis_staging.R
\name{derive_diagnosis_staging}
\alias{derive_diagnosis_staging}
\title{Derive, label, and add diagnosis and staging variables to the merged data set.}
\usage{
derive_diagnosis_staging(data)
}
\arguments{
\item{data}{A data frame containing VMAC variables.}
}
\value{
\code{data} with added diagnosis and staging variables.
}
\description{
Derive, label, and add diagnosis and staging variables to the merged data set.
}
