% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derive_family_hx.R
\name{derive_family_hx}
\alias{derive_family_hx}
\title{Derive, label, and add family history variables to the merged data set.}
\usage{
derive_family_hx(data)
}
\arguments{
\item{data}{A data frame containing VMAC variables.}
}
\value{
\code{data} with added family history variables.
}
\description{
Derive, label, and add family history variables to the merged data set.
}
