% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derive_food_scan.R
\name{derive_food_scan}
\alias{derive_food_scan}
\title{Derive, convert, label, and add food scan variables to the merged data set.}
\usage{
derive_food_scan(data)
}
\arguments{
\item{data}{A data frame containing VMAC variables.}
}
\value{
\code{data} with added food scan variables.
}
\description{
Derive, convert, label, and add food scan variables to the merged data set.
}
