% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derive_frailty.R
\name{derive_frailty}
\alias{derive_frailty}
\title{Derive, label, and add frailty variables to the merged data set.}
\usage{
derive_frailty(data)
}
\arguments{
\item{data}{A data frame containing VMAC variables.}
}
\value{
\code{data} with added frailty variables.
}
\description{
Derive, label, and add frailty variables to the merged data set.
}
