% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derive_fsrp.R
\name{derive_fsrp}
\alias{derive_fsrp}
\title{Derive, label, and add FSRP variables to the merged data set.}
\usage{
derive_fsrp(data)
}
\arguments{
\item{data}{A data frame containing VMAC variables.}
}
\value{
\code{data} with added FSRP variables.
}
\description{
Derive, label, and add FSRP variables to the merged data set.
}
