% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derive_informant_info.R
\name{derive_informant_info}
\alias{derive_informant_info}
\title{Derive, label, and add informant information variables to the merged data set.}
\usage{
derive_informant_info(data)
}
\arguments{
\item{data}{A data frame containing VMAC variables.}
}
\value{
\code{data} with added informant information variables.
}
\description{
Derive, label, and add informant information variables to the merged data set.
}
