% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derive_medical_hx.R
\name{derive_medical_hx}
\alias{derive_medical_hx}
\title{Derive, label, and add medical history variables to the merged data set.}
\usage{
derive_medical_hx(data)
}
\arguments{
\item{data}{A data frame containing VMAC variables.}
}
\value{
\code{data} with added medical history variables.
}
\description{
Derive, label, and add medical history variables to the merged data set.
}
