% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derive_medication_surgery.R
\name{derive_medication_surgery}
\alias{derive_medication_surgery}
\title{Derive, label, and add medication and surgery variables to the merged data set.}
\usage{
derive_medication_surgery(data, diabetes.file, cholesterol.file, afib.file,
  afib_surgery.file, AH_beta_blocker.file,
  AH_beta_blocker_if_not_drop.file, AH_ace_inhibitor.file, AH_arb.file,
  AH_ccb.file, AH_ksd.file, AH_other.file)
}
\arguments{
\item{data}{A data frame containing VMAC variables.}

\item{diabetes.file, cholesterol.file, afib.file, afib_surgery.file, AH_beta_blocker.file, AH_beta_blocker_if_not_drop.file, AH_ace_inhibitor.file, AH_arb.file, AH_ccb.file, AH_ksd.file, AH_other.file}{File paths to the respective data sets.}
}
\value{
\code{data} with added medication and surgery variables.
}
\description{
Derive, label, and add medication and surgery variables to the merged data set.
}
