% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derive_mlta_champs.R
\name{derive_mlta_champs}
\alias{derive_mlta_champs}
\title{Derive, convert, label, and add MLTA and CHAMPS variables to the merged data set.}
\usage{
derive_mlta_champs(data)
}
\arguments{
\item{data}{A data frame containing VMAC variables.}
}
\value{
\code{data} with added MLTA and CHAMPS 3T variables.
}
\description{
Derive, convert, label, and add MLTA and CHAMPS variables to the merged data set.
}
