% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derive_neuropsych.R
\name{derive_neuropsych}
\alias{derive_neuropsych}
\title{Derive, label, and add neuropsychological variables to the merged data set.}
\usage{
derive_neuropsych(data)
}
\arguments{
\item{data}{A data frame containing VMAC variables.}
}
\value{
\code{data} with added neuropsychological variables.
}
\description{
Derive, label, and add neuropsychological variables to the merged data set.
}
