% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derive_psqi.R
\name{derive_psqi}
\alias{derive_psqi}
\title{Derive, label, and add PSQI variables to the merged data set.}
\usage{
derive_psqi(data)
}
\arguments{
\item{data}{A data frame containing VMAC variables.}
}
\value{
\code{data} with added PSQI variables.
}
\description{
Derive, label, and add PSQI variables to the merged data set.
}
