% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derive_pvlt.R
\name{derive_pvlt}
\alias{derive_pvlt}
\title{Derive, label, and add PVLT variables to the merged data set.}
\usage{
derive_pvlt(data)
}
\arguments{
\item{data}{A data frame containing VMAC variables.}
}
\value{
\code{data} with added PVLT variables.
}
\description{
Derive, label, and add PVLT variables to the merged data set.
}
