% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exclude_participants.R
\name{exclude_participants}
\alias{exclude_participants}
\title{Apply exclusion criteria to a data set, ideally as part of an inclusion/exclusion section of the code.}
\usage{
exclude_participants(map_id, usable, diagnosis, predictors, outcomes,
  covariates, data, data_name = NULL, data_type)
}
\arguments{
\item{map_id}{A character vector containing one or more three-digit MAP IDs.}

\item{usable}{A character vector containing one or more "usable" variable names.}

\item{diagnosis}{A character vector containing one or more of the following values: \code{"Dementia"}, \code{"Normal"}, \code{"Ambiguous At Risk"}, and/or \code{"MCI"}. Each diagnostic group will be removed in its own step. A value of NULL for this argument will skip this step.}

\item{predictors}{A character vector containing predictor variable names. Observations with all predictors missing will be removed in one step. A value of NULL for this argument will skip this step.}

\item{outcomes}{A character vector containing outcome variable names. Observations with all outcomes missing will be removed in one step. A value of NULL for this argument will skip this step.}

\item{covariates}{A character vector containing covariate variable names. For each covariate, observations with a missing value will be removed in its own step. A value of NULL for this argument will skip this step.}

\item{data}{A data frame to which the exclusion criteria should be applied.}

\item{data_name}{Optional. By default, this will be set to the name of the object supplied in \code{data}. Supplying a value will overwrite this default behavior.}

\item{data_type}{An atomic character string of either \code{"b"} for baseline/cross-sectional data or \code{"l"} for longitudinal data.}
}
\value{
\code{data} with exclusion criteria applied.
}
\description{
Apply exclusion criteria to a data set, ideally as part of an inclusion/exclusion section of the code.
}
\examples{
exclude_participants(diagnosis = c("Dementia", "Ambiguous At Risk"), predictors = predictors.var, outocmes = outcomes.var, covariates = covariates.var, data = mydat0.long, data_type = "l")
}
