% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_id.R
\name{format_id}
\alias{format_id}
\title{Format MAP IDs to strings with three digits.}
\usage{
format_id(data)
}
\arguments{
\item{data}{A data frame that contains the variable \code{map_id} or \code{map.id}.}
}
\value{
\code{data} that contains the formatted variable \code{map_id} or \code{map.id}.
}
\description{
Format MAP IDs to strings with three digits.
}
