% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_names.R
\name{format_names}
\alias{format_names}
\title{Replace underscores with periods in a data frame's variable names.}
\usage{
format_names(data)
}
\arguments{
\item{data}{A data frame.}
}
\value{
\code{data} with underscores replaced with periods in the variable names.
}
\description{
Replace underscores with periods in a data frame's variable names.
}
