% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_p.R
\name{format_p}
\alias{format_p}
\title{Format a vector of p-values to be colored red if less than the specified alpha.}
\usage{
format_p(vec, alpha = 0.05)
}
\arguments{
\item{vec}{A numeric vector containing p-values.}

\item{alpha}{A numeric significance value.}
}
\value{
A \code{formattable} object containing the formatted vector of p-values.
}
\description{
Format a vector of p-values to be colored red if less than the specified alpha.
}
\examples{
\dontrun{
formattable::formattable(summary.all.df, list(`p-Value` = formatP, `Total p-Value` = formatP))
}
}
