% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exclude_participants.R
\name{generate_inclusion_exclusion}
\alias{generate_inclusion_exclusion}
\title{Generate a data frame that summarizes the inclusion/exclusion criteria. This function is not intended to be explicitly called by the user.}
\usage{
generate_inclusion_exclusion(step, map_id = NULL, diagnosis = NULL,
  vars = NULL, data, data_name, data_type)
}
\arguments{
\item{step}{An atomic character string from the following values: \code{"first"}, \code{"map_id"}, \code{"usable"}, \code{"diagnosis"}, \code{"predictor"}, \code{"outcome"}, \code{"covariate"}, \code{"final"}.}

\item{map_id}{Optional. Supply only if the argument step is set to "map_id." A character vector containing one or more three-digit MAP IDs.}

\item{diagnosis}{Optional. Supply only if the argument step is set to "diagnosis." An atomic character string containing one of the following values: \code{"Dementia"}, \code{"Normal"}, \code{"Ambiguous At Risk"}, \code{"MCI"}.}

\item{vars}{Optional. Supply only if the argument step is set to "usable", "predictor," "outcome," or "covariate." A character vector containing the variable or variables to check for missing values.}

\item{data}{A data frame to which the exclusion criteria is applied.}

\item{data_name}{The name of the data object to print in the output.}

\item{data_type}{An atomic character string of either \code{"b"} for baseline/cross-sectional data or \code{"l"} for longitudinal data.}
}
\value{
A data frame named \code{flow_chart.df} containing a summary of the exclusion process.
}
\description{
Generate a data frame that summarizes the inclusion/exclusion criteria. This function is not intended to be explicitly called by the user.
}
