% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invalidate_color_blind.R
\name{invalidate_color_blind}
\alias{invalidate_color_blind}
\title{Invalidate certain variables due to color blindness in certain participants in the merged data set.}
\usage{
invalidate_color_blind(data)
}
\arguments{
\item{data}{A data frame containing VMAC variables.}
}
\value{
\code{data} with invalidated color blindness-related variables.
}
\description{
Invalidate certain variables due to color blindness in certain participants in the merged data set.
}
