% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invalidate_derived_neuropsych_items.R
\name{invalidate_derived_neuropsych_items}
\alias{invalidate_derived_neuropsych_items}
\title{Invalidate certain neuropsychological variables for certain participants in the merged data set.}
\usage{
invalidate_derived_neuropsych_items(data)
}
\arguments{
\item{data}{A data frame containing VMAC variables.}
}
\value{
\code{data} with invalidated neuropsychological variables.
}
\description{
Invalidate certain neuropsychological variables for certain participants in the merged data set.
}
