% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invalidate_neuropsych_addendum.R
\name{invalidate_neuropsych_addendum}
\alias{invalidate_neuropsych_addendum}
\title{Invalidate certain neuropsychological addendum variables for certain participants in the merged data set.}
\usage{
invalidate_neuropsych_addendum(data)
}
\arguments{
\item{data}{A data frame containing VMAC variables.}
}
\value{
\code{data} with invalidated neuropsychological addendum variables.
}
\description{
Invalidate certain neuropsychological addendum variables for certain participants in the merged data set.
}
