% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invalidate_neuropsych_eligibility.R
\name{invalidate_neuropsych_eligibility}
\alias{invalidate_neuropsych_eligibility}
\title{Invalidate certain neuropsychological eligibility variables for certain participants in the post-fixed eligibility data set.}
\usage{
invalidate_neuropsych_eligibility(data)
}
\arguments{
\item{data}{A data frame containing VMAC variables.}
}
\value{
\code{data} with invalidated neuropsychological eligibility variables.
}
\description{
Invalidate certain neuropsychological eligibility variables for certain participants in the post-fixed eligibility data set.
}
