% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_pkg.R
\name{load_pkg}
\alias{load_pkg}
\title{Load multiple packages, installing them from CRAN first if they are not already installed.}
\usage{
load_pkg(pkg)
}
\arguments{
\item{pkg}{A character vector containing the names of local or CRAN packages.}
}
\description{
Load multiple packages, installing them from CRAN first if they are not already installed.
}
