% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_within_epoch.R
\name{merge_within_epoch}
\alias{merge_within_epoch}
\title{Major function to merge raw, non-derived data within each epoch.}
\usage{
merge_within_epoch(epoch, save.file = NULL, main, abp = NULL,
  biomarkers = NULL, auto3T = NULL, auto3TBH = NULL, man3T = NULL,
  man3TBH = NULL, qmass = NULL, addendum = NULL, csf = NULL,
  srt = NULL, breathhold.prefix = "bHold.")
}
\arguments{
\item{epoch}{Epoch number}

\item{save.file}{An optional file path to an RDS object where the merged data will be saved.}

\item{main}{MAP Main data}

\item{abp}{MAP ABP Consent data}

\item{biomarkers}{MAP Biomarkers data}

\item{auto3T}{MAP Automated 3T Brain MRI data}

\item{auto3TBH}{MAP Automated 3T Brain MRI Breath Hold data}

\item{man3T}{MAP Manual 3T Brain MRI data}

\item{man3TBH}{MAP Manual 3T Brain MRI Breath Hold data}

\item{qmass}{MAP Cardiac MRI Analysis data}

\item{addendum}{MAP Addendum data}

\item{csf}{MAP Clinical CSF data}

\item{srt}{MAP SRT Error Analysis data}

\item{breathhold.prefix}{A string containing the temporary prefix to be added to the breath hold variables.}
}
\value{
A data frame with the above files merged together.
}
\description{
Major function to merge raw, non-derived data within each epoch.
}
