% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_excluded_participants.R
\name{print_excluded_participants}
\alias{print_excluded_participants}
\title{Use in statistical reports to print MAP IDs excluded from a specific exclusion step.}
\usage{
print_excluded_participants(step)
}
\arguments{
\item{step}{An integer specifying the exclusion step number (stored in a list called \code{id.excl}).}
}
\value{
A sentence stating that no participants were excluded in the specified step or a sentence with a comma separated list of excluded participants.
}
\description{
Use in statistical reports to print MAP IDs excluded from a specific exclusion step.
}
\examples{
excl_map_id(step = 2)
}
