% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_calculated_fields.R
\name{process_calculated_fields}
\alias{process_calculated_fields}
\title{Invalidation of REDCap calculated fields.}
\usage{
process_calculated_fields(data, data_label, epoch = current_epoch)
}
\arguments{
\item{data}{A dataframe containing data from REDCap.}

\item{data_label}{The label corresponding to the data as stored in MAPfreeze.list (used to locate metadata).}

\item{epoch}{An integer corresponding to the time period of data (used to locate metadata).}
}
\value{
The input dataframe with properly invalidated observations in REDCap calculated variables.
}
\description{
Numeric observations in VMAP data are coded as invalid with negative multiples of 1111 (e.g. -8888, -9999, etc.). While these observations are converted to NA using \code{missing_to_na} in the \code{process_raw_data} script, \code{missing_to_na} is unable to always detect invalid values in calculated fields from REDCap. \code{process_calculated_fields} uses regular expressions to extract which variables were utilized in the REDCap calculation of a given variable. If any of the component variables were invalidated by \code{missing_to_na}, \code{process_calculated_fields} will convert the corresponding calculated values to NA.
}
\examples{
apoe_processed.df <- process_calculated_fields(data = apoe.df, data_label = "APOE")
}
