% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_eligibility.R
\name{process_eligibility}
\alias{process_eligibility}
\title{Apply various processing operations to the eligibility data set.}
\usage{
process_eligibility(elig_data)
}
\arguments{
\item{elig_data}{A data frame containing VMAC eligibility data.}
}
\value{
A processed version of \code{elig_data}.
}
\description{
Apply various processing operations to the eligibility data set.
}
