% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_raw_data.R
\name{process_raw_data}
\alias{process_raw_data}
\title{Major function to process raw, non-derived data within each epoch.}
\usage{
process_raw_data(epoch, main, abp = NULL, biomarkers = NULL,
  auto3T = NULL, auto3TBH = NULL, man3T = NULL, man3TBH = NULL,
  qmass = NULL, addendum = NULL, csf = NULL, srt = NULL,
  breathhold.prefix = "bHold_")
}
\arguments{
\item{epoch}{Epoch number}

\item{main}{MAP Main data}

\item{abp}{MAP ABP Consent data}

\item{biomarkers}{MAP Biomarkers data}

\item{auto3T}{MAP Automated 3T Brain MRI data}

\item{auto3TBH}{MAP Automated 3T Brain MRI Breath Hold data}

\item{man3T}{MAP Manual 3T Brain MRI data}

\item{man3TBH}{MAP Manual 3T Brain MRI Breath Hold data}

\item{qmass}{MAP Cardiac MRI Analysis data}

\item{addendum}{MAP Addendum data}

\item{csf}{MAP Clinical CSF data}

\item{srt}{MAP SRT Error Analysis data}

\item{breathhold.prefix}{A string containing the temporary prefix to be added to the breath hold variables.}
}
\value{
MAPfreeze.list is updated with the processed data.
}
\description{
Major function to process raw, non-derived data within each epoch.
}
