% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proportion_non_missing.R
\name{proportion_non_missing}
\alias{proportion_non_missing}
\title{Calculate the proportion of non-missing values in a vector.}
\usage{
proportion_non_missing(vec)
}
\arguments{
\item{vec}{A vector.}
}
\value{
A number between 0 and 1 indicating the proportion of non-missing values in \code{vec}.
}
\description{
Calculate the proportion of non-missing values in a vector.
}
\seealso{
Other scoring functions: \code{\link{reverse}},
  \code{\link{total_score}}
}
\concept{scoring functions}
