% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reverse_score.R
\name{reverse}
\alias{reverse}
\alias{reverse0to1}
\alias{reverse1to3}
\alias{reverse1to7}
\title{Reverse the scores in a vector by subtraction.}
\usage{
reverse0to1(vec)

reverse1to3(vec)

reverse1to7(vec)
}
\arguments{
\item{vec}{A numeric vector.}
}
\value{
\code{vec} with values reversed, based on which function was used.
}
\description{
Reverse the scores in a vector by subtraction.
}
\section{Functions}{
\itemize{
\item \code{reverse0to1}: Reverse the scores in a vector that normally range from 0 to 1.

\item \code{reverse1to3}: Reverse the scores in a vector that normally range from 1 to 3.

\item \code{reverse1to7}: Reverse the scores in a vector that normally range from 1 to 7.
}}

\seealso{
Other scoring functions: \code{\link{proportion_non_missing}},
  \code{\link{total_score}}
}
\concept{scoring functions}
