% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/round_to.R
\name{round_to}
\alias{round_to}
\title{Round the elements of a numeric vector to a specified decimal place.}
\usage{
round_to(x, k)
}
\arguments{
\item{x}{A numeric vector.}

\item{k}{An integer indicating the number of decimals to show.}
}
\value{
A character vector of length as \code{x} containing the rounded numbers.
}
\description{
Round the elements of a numeric vector to a specified decimal place.
}
\examples{
round_to(pi, 5)
round_to(c(exp(1), 0, 1.3), 2)

\dontrun{
round_to("pi", 5)
}
}
