% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/total_average_score.R
\name{total_score}
\alias{total_score}
\alias{average_score}
\title{Calculate the total (sum) or the average (mean) score of a numeric vector.}
\usage{
total_score(vec, threshold = 0.85, digits = 1)

average_score(vec, threshold = 0.85, digits = 1)
}
\arguments{
\item{vec}{A numeric vector.}

\item{threshold}{A number between 0 and 1 that indicates the minimum acceptable proportion of non-missing values in \code{vec}.}

\item{digits}{A non-negative integer specifying the number of digits to round the score to.}
}
\value{
The sum or mean of the numbers contained in \code{vec} rounded to \code{round} digits. If the proportion of non-missing values in \code{vec} is less than \code{threshhold}, the functions will return NA.
}
\description{
Calculate the total (sum) or the average (mean) score of a numeric vector.
}
\examples{
average_score(c(1, 2, NA, 4))
}
\seealso{
Other scoring functions: \code{\link{proportion_non_missing}},
  \code{\link{reverse}}
}
\concept{scoring functions}
